init/#1: move package
